Keyup events in the console could cancel replay.
The follwoing scenario:
 - record a replay with spacebar held down for a long period
 - replay this recording
 - while replaying, open the console and press space

Because of a bug in the console code, the keyup event was send to the MSX and
the replay was canceled. See comments in code for more details (bug was caused
by SDL not filling in the unicode field for keyup events).

Now the console again consumes all keyup events (was also like this before
revision 11232).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11265 7512d703-071c-0410-b5c2-d2bdd895b7d1