Another fix for hanging keys with reverse.
After a lot of searching I could reliably reproduce it like this:
 - boot openmsx to basic
 - press 'a', and keep it pressed
 - once the (msx-basic) auto-repeat kicks in, press PgUp (this goes back to
   a situation where 'a' was not yet pressed)
 - release PgUp and almost immediately after also release 'a'
(If someone files another bug report on this .. it REALLY helps if you can
provide a scenario like the one above to debug this.)

Before this fix, the 'a' key was stuck, now it seems to work fine.

I'm not 100% sure, but I think that I still got a hanging key once after this
fix. Though I can't reproduce it (and from the debug prints I added, it seemed
that the keyup event was really missing (not somehow ignored by the keyboard
code)).

So we should still keep an eye on this. But in any case it's certainly better
with this patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11285 7512d703-071c-0410-b5c2-d2bdd895b7d1