Fixed another part of the 'hanging keys on reverse' bug.
The EventDelay class can (for a short time) delay MSX input events. If such an
event was scheduled to be processed in the (near) future, but before that
moment in time is reached, we perform a 'reverse goto', then this event was
lost. But to keep the host/msx keyboard synchronization in sync, it's important
that the MSX keybaord code doesn't miss any events.

Fixed by on 'reverse goto' flush all pending events (execute them immediately).
The (relative) timing of these events is not correct anymore, but at least none
get lost anymore. (Another unsuccessful attempt to fix this tried to reschedule
the events in the new msx machine).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11304 7512d703-071c-0410-b5c2-d2bdd895b7d1