Add -mono and -stereo options to record
By default, a stereo recording will be made if there are any stereo
devices present or a device has its balance off-center. If a mono 
recording is started (either because -mono is specified or stereo is
not required at the time the recording was started) and stereo is found,
a warning is given and the channels will be mixed down to mono.


git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11316 7512d703-071c-0410-b5c2-d2bdd895b7d1