Fix [2961441] Diskmanipulator crash while importing
There were actually two bugs:
1) the class MSXtar was not robust enough against invalid disk formats
2) the DiskImageUtils::format() function generated invalidly formatted disks
   for certain disk sizes (e.g. for a 2MB disk size)

The problem was that 2) generated a disk image with too few FAT sectors to be
able to hold all the clusters in the disk image. The MSXtar class 1) calculated
the maximum cluster number from the number of sectors and thus happily wrote
past the end of the FAT.

I added an extra check in MSXtar to limit the maximum cluster number to the
maximum capacity of the FAT. So effectively a (big) part of the image becomes
useless now, but it's anyway an invalid disk format.

I also fixed 2) so that it allocates enough FAT sectors for all the possible
image sizes.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11317 7512d703-071c-0410-b5c2-d2bdd895b7d1