Another fix for the hanging-key-on-reverse problem (hopefully final fix now)
There were still some problems with combinations of pressing/releasing keys
right before/after going back in time, if those keys were located in the same
row of the MSX keyboard matrix (see bug [2922518] for more detailed scenarios).

The underlying cause of these problems was the following:
 a) keyboard code receives a keyboard event
 b) it figures out if and how this event changes the MSX keyboard matrix
 c) if replay was still active, then deactivate it
 d) apply the change to the MSX keyb matrix
In this sequence, step c) changes the msx keyb matrix state. But that state was
already used in step b) to calculate the new state for (one row) in the keyb
matrix.

Fixed by splitting step b) in two. First figure out if the event changes the
MSX keyb state. If so then stop replay (step c). And only then use the values
in the keyb matrix to calculate the new keyb matrix state.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11324 7512d703-071c-0410-b5c2-d2bdd895b7d1