Fix interrupt and revert bogus fix for Rolling Blaster
When the video stops and superimposing is enabled, an interrupt is 
generated. If superimposing is not enabled, register EXTINT will
still read true and an interrupt will be generated if superimposing
is enabled at a later stage before EXTINT is read.

Note that Rolling Blaster will again not work, but the previous fix was
not correct. Now Cosmic Circuit and Starfighters will work again.


git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11334 7512d703-071c-0410-b5c2-d2bdd895b7d1