Fixed crash when the Tcl command on a breakpoint gave an error
Could be reproduced like this:
  proc foo {} { bar }
  debug set_bp 0x0038 true foo
  reverse goto 0.123
With reverse enabled and with 'bar' an invalid command. Was originally
discovered with Quibus' dostrace.tcl script.

When the Tcl command associated with a breakpoint (or watchpoint, ...)
generates an error, that error is printed via the CliComm printWarning
mechanism. The problem was that before this patch, breakpoint objects held a
reference to the MSX machine specific CliComm object. And after a reverse (or
loadstate) we're using a different MSX machine and thus this CliComm object was
already deleted. Fixed by requiring the global CliComm object for breakpoints.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11371 7512d703-071c-0410-b5c2-d2bdd895b7d1