Fixed audio/video sync problem in mono avi recordings
The problem was caused by rounding errors. As a left-over from the stereo-only
recording code, the number of samples were divided by 2 (the get the number of
fragments) and later again multiplied by two (to get again number of samples).
This is OK for stereo (there always are an even number of samples), but for
mono this sometimes looses one sample. This effect exists in both PAL/NTSC
recordings, but some reason it was much easier to notice in PAL.

Thanks to Manuel and Maarten for doing to hard work of figuring out what was
wrong, I merely implemented the fix.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11390 7512d703-071c-0410-b5c2-d2bdd895b7d1