Fixed EventDelay implementation
Actually the bug was there since the original implementation (6 years ago!).
The bug could cause events to be processed too far in the future. This effect
was small, but it became very visible with the reverse/replay feature. Because
after a long replay (actually a long 'reverse goto') there's also a long time
bewteen EventDelay::sync() calls and the event scheduling bug becomes worse.
Even so bad that events can get scheduled out-of-order, and this can cause
weird effects like (seemingly) keyup/down swapped or keyb reacts to the
previous instead of the last pressed key.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11413 7512d703-071c-0410-b5c2-d2bdd895b7d1