Fixed problem with reverse/replay and cassetteplayer.
If replay is enabled from MSX time zero (e.g. when auto_enable_reverse is
active). And you go back to the very first snapshot (either with 'reverse goto
0' or by saving/loading the replay log). Then the cassette player could have
lost the inserted cassette image.

This problem was caused by the order of the auto-enable-reverse and the auto-
plug-cassette-player scripts. If auto-plug runs first it's ok, in the other
order it causes the above problem. The order in which the startup scripts are
executed is random, so this is a bug.

See the comments in this patch for a more detailed explanation of what goes
wrong in the above scenario and how it's fixed (for now).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11418 7512d703-071c-0410-b5c2-d2bdd895b7d1