Also update FAT2 in updateFileInDisk()
Before this change it could happen that after using dir-as-disk, some host
files got updated modification timestamps (so they were written to). Even if
those files were not accessed (neither read nor written) by the MSX.

The reason why this happened is that while initially importing the host files
in the virtual disk image, the dir-as-disk code only updates FAT1. Later when
the MSX writes to a FAT sector (e.g. when you create or extend a file from the
MSX side), FAT1 and FAT2 are compared for differences, and the files that
correspond to those FAT differences are exported to the host filesystem. So
(some of) the initially imported files could needlessly be re-exported.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11428 7512d703-071c-0410-b5c2-d2bdd895b7d1