Improved error handling in updateFileInDisk()
In case of an error (virtual MSX disk image full, or host read error) we
should still properly terminate the FAT chain in the virtual disk image.

In case of zero-sized MSX files (either because of an error or simply because
of a zero-sized host file), we shouldn't allocate any cluster (checked on a
MSXturboR, DOS2).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11429 7512d703-071c-0410-b5c2-d2bdd895b7d1