Fixed dir-as-disk corruption with an (almost) full diskimage.
Before this patch, the maximum allowed cluster number was 720, while it should
be only 714 for a normal 1440 sector DD disk. When the virtual disk image was
almost full (when it starts using clusters 715-720), it also starts using
sector numbers 1440-1451. But the code uses a fixed-size array of size 1440 to
store meta information about sectors. These out-of-array-bound writes overflow
into the virtual FAT area which can later even result in corrupt host files!

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11430 7512d703-071c-0410-b5c2-d2bdd895b7d1