Change BlipBuffer interface to use delta instead of absolute amplitudes.
On it's own this is already a (very) minor optimization (no need to store
previous amplitude and calculate deltas internally). But this is also a
preparation to in the future let SoundDevices (optionally) generate a sparse
streams of deltas instead of a dense stream of absolute amplitudes (useful for
PSG and SCC).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11512 7512d703-071c-0410-b5c2-d2bdd895b7d1