Added various VDP timing related info topics:
 - VDP_frame_count
 - VDP_cycle_in_frame
 - VDP_line_in_frame
 - VDP_cycle_in_line
 - VDP_msx_y_pos
 - VDP_msx_x256_pos
 - VDP_msx_x512_pos

Most important is 'VDP_frame_count'. Before this patch we estimated frame count
by dividing current time by the duration of one frame, but this doesn't work if
the framerate can be switched (PAL/NTSC).

The other info topics can be used to query the current scanline position. It
can be queried in absolute numbers (in lines and cycles) or sometimes more
convenient in 'MSX coordinates'.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11541 7512d703-071c-0410-b5c2-d2bdd895b7d1