Implemented 'disablesprites' setting and 'screenshot -no-sprites' command
Added a new boolean setting 'disablesprites' that can disable sprite rendering
for both V99x8 and V9990. Only the rendering itself is disabled, stuff like
sprite collision detection or VDP command engine speed is left unchanged (so
the MSX can't notice the state of this setting).

Removed BiFi's 'screencapture.tcl' script. Instead the screenshot command
itself is extended with a new '-no-sprites' option. The actual implementation
of this option is still in Tcl (and uses the new 'disablesprites' setting). It
works by renaming the built-in openmsx command and replacing it with a Tcl proc
(which delegates to the built-in command).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11545 7512d703-071c-0410-b5c2-d2bdd895b7d1