Calculate sequence numbers for reverse snapshots
Before this patch we kept a sequence number as member variable and increased
that by one each time a snapshot was taken. Now we always calculate the
sequence number based on the difference in EmuTime bewteen the time of the
first snapshot and the current time.

Normally both methods give the same result. But when performing jumps in time
there is a difference. For example when you load a long replay and immediately
jump to the end, the old method would give seq num 2 to the new snapshot, while
the new method uses a much higher seq number (which is correct because if you
go back to somewhere in between the seq numbers should also lie in between).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11550 7512d703-071c-0410-b5c2-d2bdd895b7d1