Added a sanity assert and a fix, because it went off...
When the last snapshot was exactly on the edge of the time line, it
wasn't included in the replay. Fixed this by improving the length of the
time line which is used for partitioning. The end-time is now not
determined anymore by the last snapshot's time, but by the last event's
time, via getEndTime. Also and most importantly, we add a tiny bit of
time to the total length to make sure we don't skip the last snapshot.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11560 7512d703-071c-0410-b5c2-d2bdd895b7d1