Use clock_gettime() instead of gettimeofday()
The problem with gettimeofday() is that it's not guarateed to be monotonic
(= never go backwards in time, even when e.g. the user or a ntp service adjusts
the time). Manuel triggered an assert because of this problem.

The function clock_gettime() with CLOCK_MONOTONIC as clock-id parameter is
guaranteed to be monotonic. Though it's not as widely available as
gettimeofday(). For example mingw has gettimeofday() but not the
CLOCK_MONOTONIC option for clock_gettime() (it does have clock_gettime()
itself). Luckily it's possible to test whether CLOCK_MONOTONIC is supported. If
it's not available we fall back to SDL_GetTicks() which is also monotonic but
has less precision.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11565 7512d703-071c-0410-b5c2-d2bdd895b7d1