Optimization in FilePool sha1sum caching
Trust that when the modification time of a file hasn't changed, the cached
sha1sum is also still valid. Before this patch we already used this assumption
to avoid having to re-scan all files in the systemroms subdirectory. But when a
match was found in the database, for extra safety, we anyway still checked the
sha1sum.

When loading a replay file, calculating sha1sums is the most expensive part.
This patch (together with the next patches) should help to reduce that.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11567 7512d703-071c-0410-b5c2-d2bdd895b7d1