Remove support for <filesize> and <skip_headerbytes>
Before this patch it was possible to inside a <rom> section in a config file to
skip to initial N bytes or only use the first M bytes (or a combination) of a
.rom file. This patch removes that possibility (it now prints an error when
these tags are still detected).

These tags were already not used for a long time in any of the official machine
or extension configs. This option also makes it more difficult to implement
sha1sum caching for rom files (will be implemented in a follow up patch). In
fact the old code already had a bug when one of these two tags was used in
combination with the <sha1> tag.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11568 7512d703-071c-0410-b5c2-d2bdd895b7d1