Speedup loading of long replay files
The repeated calls to the beginTag() method used when deserializing a XML file
(for loadstate or load replay) had a combined complexity of O(n^2). This patch
adds a findNextChild() method in XMLElement so that the combined complexity can
be reduced to O(n).

This mostly matters when loading a long replay file, because (only) there are
tags with a huge amount of direct child tags (the <items> in <events>).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11573 7512d703-071c-0410-b5c2-d2bdd895b7d1