Cleanly shutdown PreCacheFile thread
A reminder: the PreCacheFile mechansim for example reads a disk image into the
OS file cache so that when that image is located on CDROM, the CDROM doesn't
constantly need to spin up when the MSX gradually reads more data from disk.

When a file object is very short lived, it can happen that pre-caching is still
busy when the object is deleted. Before this patch the thread was abruptly
terminated. Now the main thread sets a (locked) boolean to request the helper
thread to terminate, and then actually waits for the thread to terminate.

Before this patch I sometimes got a 'terminate called without an active
exception' error and a crash when running openmsx in valgrind. And the
crashdump pointed to this PreCacheFile thread (I never saw this crash without
using valgrind). After this patch the crash is gone.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11574 7512d703-071c-0410-b5c2-d2bdd895b7d1