Allow to get file modification date without having to actually open the file
This might give a small speed up for scanning through the systemroms directory.
This commonly happens when in a config file a file is specified both with sha1
and filename and the file is actually located in the <config>/roms instead of
the systemroms directory. (Though because of revision 11569 the scanning only
happens the first this file is used).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11578 7512d703-071c-0410-b5c2-d2bdd895b7d1