Speedup Base64::encode() and Base64::decode()
Immediately set output buffer to a big enough (too big) size (and shrink to
correct size at the end). This avoids the occasional reallocations that were
needed before this patch.

An alternative implementation is to use std::string::reserve() instead of
resize(). This has the advantage that initially the buffer can be left
uninitialized (now it is filled with dummy values). But the disadvantage is
that each append() still checks whether the buffer is big enough (will always
be the case). I measured both alternatives and this was the fastest.

This results in a small speedup when loading a replay file.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11579 7512d703-071c-0410-b5c2-d2bdd895b7d1