Fixed directory creation problem with 'reverse savereplay'
When the '<openmsx-user-directory>/replays' subdirectory doesn't exist yet, the
'reverse savereplay <name>' command would fail: it's trying to save in a
non-existing subdirectory. The 'reverse savereplay' command (so without extra
name parameter) works fine, it does first create the subdirectory.

The 0.8.0 release also has this bug. Is this bug serious enough to do a 0.8.1
release sooner?

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11580 7512d703-071c-0410-b5c2-d2bdd895b7d1