Don't un(g)zip disk images twice
I noticed diskimages were always un(g)zipped twice. This is because when you
insert a diskimage into an (emulated) disk drive. We first check if the image
is a XSA disk image (this requires to uncompres the image). If that check fails
we try a normal DSK image (this again uncompreses the image).

This patch uncompreses the file only once upfront and passes this opened file
to the XSA and DSK image constructors (before it only passed the filename).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11585 7512d703-071c-0410-b5c2-d2bdd895b7d1