Fixed auto stereo (.avi/.wav) recording for mbstereo
When starting a recording, there is code thtat tries to determine wheter the
recording should be stereo or if a mono recording is sufficient. Before this
patch that code only tested if there were any stereo sound devices present. Now
it also checks if any of the (per chip) balance settings has a non-zero value.
(Per-channel balance was already covered by the stereo check (in the current
implementation)).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11589 7512d703-071c-0410-b5c2-d2bdd895b7d1