Fix 'stop replay on toggle console' problem
During a replay, if you closed the openMSX console with the 'toggle console'
command or Cmd+L on a Mac, the replay was interrupted. Reason is that the
enter,release or Cmd/L,release event is send to the MSX (console doesn't
capture it anymore) and this will interrupt the replay even though this release
event won't influence the MSX keyboard state.

See also commit message of revision 11324. This patch fixes step b) from that
description. It now correctly figures out whether a key event will have an
influence on the MSX keyboard state. If not, there's no reason to stop replay.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11599 7512d703-071c-0410-b5c2-d2bdd895b7d1