Created helper routine to parse filename arguments in Tcl commands
Used by
 - record start
 - reverse savereplay
 - screenshot
 - cassetteplayer new

'record' and 'reverse' were already fixed in previous patches. 'screenshot' did
create file in the current working directory instead of using the standard
directory in case a filename, but no directory is given. 'cassetteplayer'
didn't create the standard directory when a filename but no directory is given
(and the standard directory didn't exist yet).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11601 7512d703-071c-0410-b5c2-d2bdd895b7d1