Implemented 'osd exists <widget>' subcommand
Added a new osd subcommand to test whether a certain osd widget exists. And used
this new command to simplify the existing scripts. Before this patch we
emulated this command by something like
  if {[catch {osd info $widgetname}]} { ... }

The main reason for this (and the previous) patch is to avoid having failing
commands in the 'normal' openmsx flow. For example the reversebar callback
procs used this 'catch-osd-info' trick very frequently. This is annoying if
you're debugging a Tcl script because it also overwrites the global 'errorInfo'
Tcl variable.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11607 7512d703-071c-0410-b5c2-d2bdd895b7d1