Simplified mode dispatch in BitmapConverter/CharacterConverter
Before this patch we used a pointer-to-member-function to dispatch to the
screen-mode specific render method. Now we use a simple switch statement.

I looked at the generated code, and the switch statement is compiled into
a jump table (so an indirect jump). Also a call via a ptr-to-mem-fun is an
indirect jump. So from a performance point of view both approaches should be
very close.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11611 7512d703-071c-0410-b5c2-d2bdd895b7d1