Simplified setting of (RGB)(A) on OSD widgets
Before this patch we had an optimization that if only the alpha part of a
widget changed (and alpha had no gradient), we don't recreate the 'content' of
the image. This code was a bit tricky and it will become even more tricky when
(in one of the next patches) we add support for borders on rectangle widgets.
Also none of our existing Tcl scripts actually benefited from this
optimization, so I simply removed it.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11619 7512d703-071c-0410-b5c2-d2bdd895b7d1