Simplified OSDRectangle::create()
For rectangles with constant alpha, OSDRectangle created a surface with
alpha=255 and applied the actual alpha when drawing the surface (OSDtext does
something similar). Though this same optimization was also already done in
SDLImage. No need to do it twice. This will also simplify the implementation
of the 'border on osd rectangles' feature (in one of the next patches).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11620 7512d703-071c-0410-b5c2-d2bdd895b7d1