Added border support for OSD rectangle
Added -bordersize -relbordersize and -borderrgba properties to the OSD
rectangle widget.

This feature is meant to replace the current osd_widgets::box Tcl proc. This
widget is implemented by combining 5 normal rectangles (1 for the background
and 4 for the borders). It works very well, though on small devices like dingoo
it can be too slow when many box widgets are used (for example in the
osd_keyboard script). Having native support for box widgets should improve both
widget creation and drawing speed.

There are a few differences between the native widget and the box script:
 - Property names are different.
 - There was no -relbordersize in the script. This property is based on the
   width of the rectangle. (Maybe in the future we can extend the border size
   properties with (optional) separate X and Y sizes).
 - The script drew the border on top of the background, while the native widget
   does not. In other words it only draws the background color in the area
   that's not covered by the border. This makes a difference when the border is
   (partly) transparent or when the background is a gradient (for the script
   the gradient starts at the corners of the whole rectangle while for the
   native widget the gradient starts inside the border).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11623 7512d703-071c-0410-b5c2-d2bdd895b7d1