Fixed bug in new rectangle border stuff:
We had an optimization to not draw fully transparent rectangles (those are
useful as parent widgets, so they are often used). Though when the background
is transparent but the border is not (and doesn't have size 0), we should of
course still draw the rectangle. This check wasn't updated yet to take the
border into account.

Side-note: for very large rectangles with transparent background but
non-transparent border (and non-zero-size border) it may be more efficient to
split the border in 4 rectangles (like we did before the native border
support). Because the current implementation always draws all rectangle pixels,
so also the interior fully transparent ones.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11630 7512d703-071c-0410-b5c2-d2bdd895b7d1