Use new rectangle border stuff for reversebar
Before we had one base rectangle with 4 narrow rectangles around it that form
the border. Now those are replaces by one bigger rectangle that contains both
background and border and one invisible rectangle that has the same size as the
background (so without border). The advantage of that invisible rectangle is
that the relative position of the other elements (progress bar and tick marks)
is still correct and that these elements can be clipped to this invisible
rectangle (so they are not drawn over the border).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11632 7512d703-071c-0410-b5c2-d2bdd895b7d1