On R800, right after a 'EI' instruction there's never a refresh.
Based on measurements on a real machine. See also doc/r800-ei-timing.ods.
TODO I still need to document this test and the results better.

From an implementation pov: I moved the R800Refresh() call from the beginning
of the instruction (common for all instructions) to the end (can be specific
for each instruction). This allowed to remove it for the 'EI' instruction.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11642 7512d703-071c-0410-b5c2-d2bdd895b7d1