Improvements to R800 'R'-register handling
On R800 the 'R'-register is a 8-bit counter (on Z80 the upper bit remains
constant while counting). Also different from Z80 is that R also increases
by one on a r800-refresh-cycle.

This count-on-r800-refresh even allowed to investigate the refresh stuff in
more detail and even already lead to improvements in the emulation of this
feature. In a few words: r800-refresh seems to wait for an even clock cycle,
see doc/r800-refresh.txt for details. This work is also still ongoing.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11643 7512d703-071c-0410-b5c2-d2bdd895b7d1