Block CTRL-<x> key presses in the console for the MSX machine
We already blocked all key up events from reaching the MSX machine when the
console was open. And we only blocked some of the CTRL-<x> key down events (the
ones that are actually used by the console). So this could lead to hanging keys
on the MSX side (e.g. press CTRL-Q when the openMSX console is opened). Fixed
by blocking all CTRL-<x> key down events when the console is open.

See also commit messages of revision 11265 and 11232.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11648 7512d703-071c-0410-b5c2-d2bdd895b7d1