Tweak LZO compression function
The LZO compression function keeps a hash table with locations of possible
matches. Before this patch, this table contained pointers, now it encodes those
pointers as offsets from the beginning of the input data.

This table is zero-initialized, so this means all hash table entries now point
to the start of the input data. This property allowed to simplify some check in
the compression loop. There was a check that the hash-entry didn't point before
or after the already processed range.

The same effect can be obtained by not changing the hash table from pointers to
offsets, but instead initialize each entry with a pointer to the start of the
input instead of a NULL pointer. I didn't do this because zero-initializing is
simpler to implement in c++, but also because it cuts the size of the table in
half on 64-bit platforms.

IIRC the original LZO code didn't initialize the has table at all, and if
that's the case then both checks were required for correctness. You gain a
little speed for not having to initialize the table, but pay for it with more
checks in the compression loop. Also not initializing the table causes some
"conditional jump depends on uninitialized value" errors in valgrind (IIRC this
was why we already did initialize the table). In rare cases this could even
lead to different compression decisions (so same input does not always compress
to the same output), though I believe in all cases the result was valid (it
still decompresses to the same output).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11651 7512d703-071c-0410-b5c2-d2bdd895b7d1