Changed LZO compression algo: faster but slightly worse compression
From a high level the LZO compression algo works like this:
  - calculate a (14-bit) hash value for the next few to-be-compressed bytes
  - use this hash value to lookup a pointer and check if there is a match
    at that location (match meaning a series of bytes that's identical
    to the current to-be-compressed bytes)
  - update the entry in the hash-table with the current position

Actually this is how the algo works after this patch. The original algorithm
also had some logic/heuristics that tried to split the hash table in a part
(full table, 14 bits) for possibly 'better' matches and a part (11 bits, so
overlaps with previous part) that works like the description above. (See
original code for details).

I did some tests and found that this heuristic did improve compression ratio a
bit (mostly between 1% and 3%, but I also had a few cases which were a tiny bit
worse with this heuristic). But removing this heuristic makes a big difference
in compression speed. The compression function itself is now about 30% faster.
And (with reverse enabled) compression takes a significant part of all CPU time
(with renderer=none and sounddriver=null, it takes 30% of all CPU time). So
IMHO speeding up compression is important even at a small cost in compression
efficiency.

This change also enables some of the cleanups in the next patches.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11652 7512d703-071c-0410-b5c2-d2bdd895b7d1