Another LZO cleanup
Removed last goto from the LZO compression function. This is actually a quite
simple change, but it changes the indentation level of a large part of the
function, so in changed lines it's rather big (that's why I committed it in
a seperate patch instead of combined with the previous one).

Also factored out common tail part in then and else branch (in a new if
introduced by previous item). And in turn moved this code to the condition of
the outer loop. This was a 'while (true) { .. if (..) break }'-loop, now it's a
simple do-while-loop.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11654 7512d703-071c-0410-b5c2-d2bdd895b7d1