Fixed security problem when replaying commands from a replay file.
It's possible someone makes a (manually edited) replay file that contains
arbitrary Tcl command in the event log. For security it's crucial that we don't
blindly execute such commands on replay.

On replay we now check that the to-be-executed command is a command that would
be recorded in the first place. So commands like these are now blocked:
  exec rm *
  after time 0 {exec rm *}
  debug set_bp 0x0038 true {exec rm *}

Actually also without this patch none of these commands would be executed on
replay. The first two not because it are not 'RecordedCommand's. The latter
would start to execute, but would then crash. This crash was not intentional,
but a lucky(?) side effect. I also can't think of an example that would
actually allow to execute arbitrary code on replay without this patch.

So while I think this patch does not fix a currently exploitable security
problem, it's of course still better to fix the code before it does become
exploitable.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11656 7512d703-071c-0410-b5c2-d2bdd895b7d1