Refine the event-listener priority system.
More specifically allow a high priority event listener to block an event from
specific lower priority listeners. Before this patch, it could only block it
for all lower priority listeners.

Used this feature to improve the event handling in the openMSX console. In
revision 11648 we blocked CTRL-<x> key events from the MSX (when the console is
open), but this had the annoying side effect that it also blocked the event for
hotkeys. So more specifically, when the console was open, it was no longer
possible to quit openMSX with the CTRL-BREAK shortcut. Should be fixed now.
This also allowed to simplify some of the other hacks we had to implement to
make the console, hotkeys and MSX subsystems work properly together.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11660 7512d703-071c-0410-b5c2-d2bdd895b7d1