Made key binding for pgup/pgdown common between tas and normal mode.
Before this patch in normal mode, pgup/down would step forward/backward
approximately one second. More specific, it takes a step to the nearest
snapshot. Also when you go back further in time, the snapshots are less dense,
so the steps get progressively bigger (with an upper limit of 15 seconds).
While in tas mode the steps were always exactly one second.

IMHO there's no reason why tas and normal mode should behave different in this
aspect. We should pick one behaviour and use it for both modes. This patch only
removes the change of key binding in tas mode. We should still decide which of
the two behaviours we prefer for the common behaviour. (There was also a bug
that leaving tas mode didn't restore the old pgup/down key bindings, that bug
disappears now).

Historically, the progressively bigger steps were useful to go back very far in
time. But now that we have a clickable reversebar, this feature is no more or
at least less useful.

Another (historic) reason for the current normal-mode pgup/down behaviour was
speed: it's faster to go to a snapshot than to some arbitrary point in between
two snapshots (and initially it was also only possible to go to an exact
snapshot position). Though in practice this speed advantage is not noticeable,
also not when the snapshots are very sparse (because we anyway limit the
maximum step size to 15 seconds).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11672 7512d703-071c-0410-b5c2-d2bdd895b7d1