Move reverse-goto black-screen-compensation from Tcl to c++
When you go back/forward in time, we actually go back 2 frame further and then
replay those two frames. This makes sure that we can display the video image
for the target time. Before this patch this trick was implemented in Tcl, now
it's moved to c++. The advantage is that it will now work for any proc that
moves in time. And it also seems to work a bit smoother.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11675 7512d703-071c-0410-b5c2-d2bdd895b7d1