Fixed V9990 commands when NX=0 or NY=0
According to the V9990 datasheet, NX=0 actually specifies a width of 2048
pixels, and NY=0 actually specifies a height of 4096 pixels. I didn't check
the current implementation on a real V9990 (e.g. what happens when NX=1500,
but the image width of the current mode is only 1024). But it's certainly
already better than before this patch.

Thanks to GuyveR800 for discovering and n_n for reporting this bug.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11690 7512d703-071c-0410-b5c2-d2bdd895b7d1