Store the current time in a replay file, on loadreplay allow to go back to that time.
Syntax for the 'reverse loadreplay' is now extended to:
   reverse loadreplay [-goto <begin|end|savetime|<n>>] <replay-file>
When not specified '-goto begin' is the default (this keeps the command
backwards compatible).

The motivation for this change was this report on IRC:
	<adelikat> I'm pressing f8 at 45385
	<adelikat> I unpaused until 45477
	<adelikat> I press f6
	<adelikat> it goes to frame 47932
	<adelikat> it should go to 45385
I can reproduce this in the case we're not at the end of the timeline when the
replay is saved (I can imagine this is a common scenario in tas mode). After
this patch the above scenario should work as expected.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11694 7512d703-071c-0410-b5c2-d2bdd895b7d1