Fixed some details in Manbow2 and MatraInk mappers
RomManbow2 and RomManbow2 both had a peek() method, but the intention was to
override the peekMem() method of the base class. Though in practice the
peekMem() implementation of the base class mostly worked fine (only inspecting
the SCC registers in the Manbow2 mapper wouldn't work).

Another detail in the Manbow2 mapper was that on power-off/-on it didn't
properly re-initialize some SCC registers (all MSXDevices that use a SCC need
to implement the powerUp() method).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11704 7512d703-071c-0410-b5c2-d2bdd895b7d1