Fix keyboard state handling after 'reverse loadreplay'
Actually make loadreplay work in the same way as reverse (see past commit
messages for many fixes related to reverse and event handling, it's quite
tricky to get right).

Consider the following scenario:
 1) boot to BASIC
 2) press the 'A' key and keep it pressed for this whole scenario,
    you will see repeating A characters on the screen
 3) go back in time (e.g. by pressing PgUp a few times)

During replay the 'A' should repeat just as they did in the original play.
Also when the end of the replay is reached, the A's should keep on repeating
(remember to keep the 'A' key pressed for the whole test).

Now a similar scenario with loadreplay:
 1 and 2) same as above
 3a) create a replay (e.g. press F8 in tas mode)
 3b) go back to that replay (e.g. press F6 in tas mode)

In this case the repeating stops as soon as the replay is loaded, even though
the A key is still being pressed. This is certainly unexpected behaviour. It
was reported as a bug by 'nitrogenesis' (a user with TAS experience).

To fix this bug, I refactored the code so that loadReplay() reuses much more
code from the goTo() method. Actually before this patch goTo() could only move
withing a single time-line, so loadReplay() first had to switch time-lines and
then use goTo() to move within that time-line. The keyboard status was not
correctly handled in the first switch. Now goTo() can also switch between
time-lines. This change also allowed to clean up the code a bit.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11718 7512d703-071c-0410-b5c2-d2bdd895b7d1